+++ modules/ensemble_dropdown/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.2
+
+* Removed deprecated methods
+
+
+## 0.1.1
+
+* Updated to Flutter 3.35.5
+* Removed deprecated methods
+
 ## 0.1.0
 
 * Minor fixes